AMP-373: add APIs and impls for workflow invocation/step
